% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtest.R
\name{ad_p_inf}
\alias{ad_p_inf}
\alias{ad_p}
\title{\eqn{\alpha} from Anderson-Darling test statistic}
\usage{
ad_p_inf(z, abs_tol = .Machine$double.eps^0.5)

ad_p(z, n, abs_tol = .Machine$double.eps^0.5)
}
\arguments{
\item{z}{the test statistic}

\item{abs_tol}{the requested accuracy of the calculation}

\item{n}{the sample size}
}
\value{
the significance level (\eqn{\alpha}).
}
\description{
Calculates the significance level (\eqn{\alpha}) from an Anderson-Darling
test statistic.
}
\details{
\code{ad_p_inf} uses the limiting distribution (i.e. for infinite sample
size). \code{ad_p} applies a correction for sample size. Since the
Anderson-Darling test statistic converges very quickly, the correction
only has a practical effect for very small sample sizes (around 5 or less).
Nevertheless, it is recommended that \code{ad_p} be used as the correction
is not computationally intensive to apply.

The method of Marsaglia and Marsaglia is used to compute the significance
level. See references below.
}
\references{
G. Marsaglia and J. Marsaglia, “Evaluating the Anderson-Darling
Distribution,” Journal of Statistical Software, vol. 9, no. 2. 25-Feb-2004.
}
