% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtest.R
\name{ad_p_unknown_param}
\alias{ad_p_unknown_param}
\title{significance level for the Anderson-Darling test statistic
when the parameters of the distribution are unknown}
\usage{
ad_p_unknown_param(z, n)
}
\arguments{
\item{z}{the test statistic}

\item{n}{the number of observations}
}
\value{
the significance level (numeric)
}
\description{
Calculates the significance level from an Anderson-Darling
test statistic. These functions assume that the parameters
of the distribution are unknown.
}
\details{
This function uses the formula for observed significance
level published in CMH-17-1G.

The results of this function have been validated against
published values in Lawless (1982).
}
\references{
“Composites Materials Handbook, Volume 1. Polymer Matrix
Composites Guideline for Characterization of Structural
Materials,” SAE International, CMH-17-1G, Mar. 2012.

J. F. Lawless, \emph{Statistical models and methods for lifetime data}.
New York: Wiley, 1982.
}
