% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_property}
\alias{analyze_property}
\title{Analyze a property as part of the analysis of a data set}
\usage{
analyze_property(an, filter, ...)
}
\arguments{
\item{an}{a \code{cmanalysis} object returned from
\code{start_analysis} or another call to
\code{analyze_property}}

\item{filter}{an data masking expression defined in terms of the variables
in the data.frame used to start the analysis.}

\item{...}{expressions used to perform the analysis or to discard
observations}
}
\description{
Once an analysis has been begun using \code{\link[=start_analysis]{start_analysis()}},
individual properties can be analyzed using this function by specifying
which observation to consider, and one or more expressions that perform
the actual analysis.
}
\details{
\code{analyze_property} takes a subset of the entire data set, and performs
one or more calculations using that subset. The subset is defined as the
observations for which \code{filter} is \code{TRUE}.

The computation(s) to perform are supplied in the \code{...} argument. One
or more function should be supplied. A \code{data.frame} will be passed to
these function as the first argument (or as an argument denoted with a dot,
for example \code{data = .}). The function(s) supplied should return objects
that have \code{glance} methods. Tidy evaluation is supported.

Normally, \code{...} will be a named (for example,
\code{b_basis = basis_normal(strength, batch)}). This name (\code{b_basis}
in this example) will be used subsequently when referring to this result.
If a name is not supplied, the name of the function called will be used.

If \code{filter} is \code{TRUE} for more than one value of the condition
variable, then all objects returned by the function(s) must be pooled across
environments. This is tested by ensuring that the \code{data.frame} returned
from calling \code{glance} on the returned object has a column named
\code{group} that contains all of the values in the condition variable (after
application of the \code{filter}.
}
\examples{
# TODO: Write this

}
\seealso{
\code{\link[=start_analysis]{start_analysis()}}
}
