% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{as.data.frame.cmanalysis}
\alias{as.data.frame.cmanalysis}
\title{Create a \code{data.frame} from an analysis}
\usage{
\method{as.data.frame}{cmanalysis}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an \code{cmanalysis} object}

\item{row.names}{not used. Included to match generic signature only}

\item{optional}{not used. Included to match generic signature only}

\item{...}{Additional parameters that are internally passed to the
\code{glance} methods.}
}
\description{
Create a \code{data.frame} based on an analysis started
using \code{\link{start_analysis}}. One row is created for
each of the combinations of test, condition and sub-group.
Internally, \code{glance} is called on each of the results
calculated using \code{\link{analyze_property}}.
}
\seealso{
\code{\link{start_analysis}}
\code{\link{analyze_property}}
}
