% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis}
\alias{basis}
\alias{basis_normal}
\alias{basis_lognormal}
\alias{basis_weibull}
\title{Calculate basis values}
\usage{
basis_normal(data = NULL, x, p = 0.9, conf = 0.95)

basis_lognormal(data = NULL, x, p = 0.9, conf = 0.95)

basis_weibull(data = NULL, x, p = 0.9, conf = 0.95)
}
\arguments{
\item{data}{a data.frame}

\item{x}{the variable in the data.frame for which to find the basis value}

\item{p}{should be 0.90 for B-Basis and 0.99 for A-Basis}

\item{conf}{confidence interval. Should be 0.95 for both A- and B-Basis}
}
\value{
an object of class \code{basis}
This object has the following fields:
\describe{
  \item{\code{call}}{the expression used to call this function}
  \item{\code{distribution}}{the distribution used (normal, etc.)}
  \item{\code{p}}{the value of \eqn{p} supplied}
  \item{\code{conf}}{the value of \eqn{conf} supplied}
  \item{\code{data}}{a copy of the data used in the calculation}
  \item{\code{n}}{the number of observations}
  \item{\code{basis}}{the basis value computed}
}
}
\description{
Calculate the basis value for a given data set. There are various functions
to calculate the basis values for different distributions. For B-Basis,
set \eqn{p=0.90} and \eqn{conf=0.95}; for A-Basis, set \eqn{p=0.90} and
\eqn{conf=0.95}
}
\details{
\code{data} is an optional argument. If \code{data} is given, it should
be a
\code{data.frame} (or similar object). When \code{data} is specified, the
value of \code{x} is expected to be a variable within \code{data}. If
\code{data} is not specified, \code{x} must be a vector.

\code{basis_normal} calculate the basis value by subtracting \eqn{k} times
the standard deviation from the mean using. \eqn{k} is given by
the function \code{\link{k_factor_normal}}.

\code{basis_lognormal} calculates the basis value in the same way
that \code{basis_normal} does, except that the natural logarithm of the
data is taken.

\code{basis_weibull} calculates the basis value for data distributed
according to a Weibull distribution. The confidence interval for the
quantile requested is calcualted using the conditional method, as
described in Lawless (1982) Section 4.1.2b. This has good agreement
with tables published in CMH-17-1G. Results differ between this function
and STAT17 by approximately 0.5%.
}
\references{
J. F. Lawless, Statistical Models and Methods for Lifetime Data.
New York: John Wiley & Sons, 1982.

“Composites Materials Handbook, Volume 1. Polymer Matrix Composites
Guideline for Characterization of Structural Materials,” SAE International,
CMH-17-1G, Mar. 2012.
}
