% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis_normal}
\alias{basis_normal}
\title{Calculate basis values (\eqn{kB}, \eqn{kA}) with normal distribution}
\usage{
basis_normal(df, x, p = 0.9, conf = 0.95)
}
\arguments{
\item{df}{a data.frame}

\item{x}{the variable in the data.frame for which to find the basis value}

\item{p}{should be 0.90 for B-Basis and 0.99 for A-Basis}

\item{conf}{confidence interval. Should be 0.95 for both A- and B-Basis}
}
\value{
an object of class \code{basis}
}
\description{
TODO: Add some text here
}
\details{
TODO: Add some text here
}
