% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon.fabric}
\alias{carbon.fabric}
\title{Sample data for a generic carbon fabric}
\format{A data frame with 216 observations and 5 variables:
\describe{
  \item{id}{a unique identification for each coupon}
  \item{test}{a code indicating the test method}
  \item{condition}{a code to indicate an environmental condition}
  \item{batch}{the prepreg batch}
  \item{strength}{the strength value observed in the test}
}}
\usage{
carbon.fabric
}
\description{
A dataset containing sample data that is typical of a generic carbon
prepreg fabric. This data is used in several examples within the
\code{cmstatr} package. This data is fictional and should
not be used for anything besides learning how to use this package.
}
\keyword{datasets}
