% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv_mean_extremum}
\alias{equiv_mean_extremum}
\title{Equivalency based on mean and extremum}
\usage{
equiv_mean_extremum(df_qual = NULL, data_qual = NULL,
  mean_qual = NULL, sd_qual = NULL, data_sample = NULL,
  n_sample = NULL, alpha, modcv = FALSE)
}
\arguments{
\item{df_qual}{(optional) a data.frame containing the qualification data.
Defaults to NULL.}

\item{data_qual}{(optional) a vector of observations from the
"qualification" data to which equivalency is being tested. Or the column of
\code{df_qual} that contains this data. Defaults to NULL}

\item{mean_qual}{(optional) the mean from the "qualification" data to which
equivalency is being tested. Defaults to NULL}

\item{sd_qual}{(optional) the standard deviation from the "qualification"
data to which equivalency is being tested. Defaults to NULL}

\item{data_sample}{(optional) a vector of observations from the sample for
which equivalency is being tested. Defaults to NULL}

\item{n_sample}{(optional) the number of observations in the sample for
which equivalency will be tested. Defaults to NULL}

\item{alpha}{the acceptable probability of a type I error}

\item{modcv}{(optional) a boolean value indicating whether a modified CV
should be used. Defaults to FALSE, in which case the standard deviation
supplied (or calculated from \code{data_qual}) will be used directly.}
}
\value{
Returns an object of class \code{equiv_mean_extremum}. This object is a list
with the following named elements:

\describe{
  \item{\code{call}}{the expression used to call this function}
  \item{\code{alpha}}{the value of alpha passed to this function}
  \item{\code{n_sample}}{the number of observations in the sample for which
    equivalency is being checked. This is either the value \code{n_sample}
    passed to this function or the length of the vector \code{data_sample}.}
  \item{\code{k1}}{the factor used to calculate the minimum individual
    threshold. The minimum individual threshold is calculated as
    \eqn{Wmin = qual_mean - k1 * qual_sd}}
  \item{\code{k2}}{the factor used to calculate the threshold for mean. The
    threshold for mean is calculated as
    \eqn{Wmean = qual_mean - k2 * qual_sd}}
  \item{\code{modcv}}{logical value indicating whether the acceptance
    thresholds are calculated using the modified CV approach}
  \item{\code{cv}}{the coefficient of variation of the qualification data.
    This value is not modified, even if \code{modcv=TRUE}}
  \item{\code{cv_star}}{The modified coefficient of variation. If
    \code{modcv=FALSE}, this will be \code{NULL}}
  \item{\code{threshold_min_indiv}}{The calculated threshold value for
    minimum individual}
  \item{\code{threshold_mean}}{The calculated threshold value for mean}
  \item{\code{result_min_indiv}}{a character vector of either "PASS" or
    "FAIL" indicating whether the data from \code{data_sample} passes the
    test for minimum individual. If \code{data_sample} was not supplied,
    this value will be \code{NULL}}
  \item{\code{result_mean}}{a character vector of either "PASS" or
    "FAIL" indicating whether the data from \code{data_sample} passes the
    test for mean. If \code{data_sample} was not supplied, this value will
    be  \code{NULL}}
  \item{\code{min_sample}}{The minimum value from the vector
    \code{data_sample}. if \code{data_sample} was not supplied, this will
    have a value of \code{NULL}}
  \item{\code{mean_sample}}{The mean value from the vector
    \code{data_sample}. If \code{data_sample} was not supplied, this will
    have a value of \code{NULL}}
}
}
\description{
Determine equivalency or equivalency thresholds based on both minimum
individual and mean.
}
\details{
There are several optional arguments to this function. However, you can't
omit all of the optional arguments. You must supply either
\code{data_sample} or \code{n_sample}, but not both. You must also supply
either \code{data_qual} (and \code{df_qual} if \code{data_qual} is a
variable name and not a vector) or both \code{mean_qual} and \code{sd_qual},
but if you supply \code{data_qual} (and possibly \code{df_qual}) you should
not supply either \code{mean_qual} or \code{sd_qual} (and visa-versa). This
function will issue a warning or error if you violate any of these rules.

If \code{modcv} is TRUE, the standard deviation used to calculate the
thresholds will be replaced with a standard deviation calculated by
multiplying \code{CV_star * mean_qual} where \code{mean_qual} is either the
value supplied or the value calculated by \code{mean(data_qual)} and
\eqn{CV* = 0.06} if \eqn{CV < 0.04}, \eqn{CV* = cv / 2 + 0.04}
if \eqn{0.04 <= cv <= 0.08} and \eqn{CV* = CV} otherwise.
}
\examples{
equiv_mean_extremum(alpha = 0.01, n_sample = 6,
                    mean_qual = 100, sd_qual = 5.5, modcv = TRUE)
#
# Call:
# equiv_mean_extremum(mean_qual = 100, sd_qual = 5.5, n_sample = 6,
#                     alpha = 0.01, modcv = TRUE)
#
# Modified CV used: CV* = 0.0675 ( CV = 0.055 )
#
# For alpha = 0.01 and n = 6
# ( k1 = 3.128346 and k2 = 1.044342 )
#               Min Individual      Sample Mean
# Thresholds:         78.88367         92.95069

}
\seealso{
\code{\link{k_equiv}}
}
