% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{hk_ext}
\alias{hk_ext}
\alias{hk_ext_z}
\alias{hk_ext_z_j_opt}
\title{Calculate values related to the Extended Hanson-Koopmans method}
\usage{
hk_ext_z(n, i, j, p, conf)

hk_ext_z_j_opt(n, p, conf)
}
\arguments{
\item{n}{the sample size}

\item{i}{the first order statistic (1 <= i < j)}

\item{j}{the second order statistic (i < j <= n)}

\item{p}{the population quantile of interest (normally 0.90 or 0.99)}

\item{conf}{the confidence bound (normally 0.95)}
}
\value{
For \code{hk_ext_z}, the return value is a numeric value representing
the parameter z (denoted as k in CMH-17-1G).

For \code{hk_ext_z_j_opt}, the return value is named list containing
\code{z} and \code{k}. The former is the value of z, as defined by
Vangel (1994), and the latter is the corresponding order statistic.
}
\description{
Calculates values related to the Extended Hanson-Koopmans method
as described by Vangel (1994).
}
\details{
Hanson (1964) presents a nonparametric method for determining
tolerance limits based on consecutive order statistics.
Vangel (1994) extends this method using non-consecutive order statistics.

The extended Hanson-Koopmans method calculates a tolerance limit
(basis value) based on two order statistics and a weighting value
\code{z}. The value of \code{z} is based on the sample size, which
order statistics are selected, the desired quantile and the desired
confidence interval.

The function \code{hk_ext_z} calculates the weighting variable \code{z}
based on selected order statistics \code{i} and \code{j}.

The function \code{hk_ext_z_j_opt} determines the value of \code{j} and
the corresponding value of \code{z}, assuming \code{i=1}. The value
of \code{j} is selected such that the computed tolerance limit is
nearest to the desired population quantile for a standard normal
distribution when the order statistics are equal to the expected
value of the order statistics for the standard normal distribution.
}
\references{
M. Vangel, “One-Sided Nonparametric Tolerance Limits,”
Communications in Statistics - Simulation and Computation,
vol. 23, no. 4. pp. 1137–1154, 1994.

D. L. Hanson and L. H. Koopmans,
“Tolerance Limits for the Class of Distributions with Increasing
Hazard Rates,” The Annals of Mathematical Statistics,
vol. 35, no. 4. pp. 1561–1570, 1964.
}
