% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{normalize}
\alias{normalize}
\title{Normalizes strength values}
\usage{
normalize(strength, measured.thk, nom.thk)
}
\arguments{
\item{strength}{the strength to be normalized. Either a vector or a numeric}

\item{measured.thk}{the measured thickness of the samples. Must be the same
length as strength}

\item{nom.thk}{the nominal thickness. Must be a single numeric value.}
}
\value{
A value (or vector) with the normalized strength values
}
\description{
This function takes a strength value (or a vector of strength values) and a
value (or vector of the same length as strength) of the measured thicknesses
and returns the normalized strength.
}
\details{
It is often necessary to normalize strength values so that variation in
specimen thickness does not unnecessarily increase variation in strength.
See CMH-17, or other references, for information about the cases where
normalization is appropriate.

Either cured ply thickness or laminate thickness may be used for
measured.thk and nom.thk, as long as CPT or laminate thickness is used
for both.

The formula applied is:
\deqn{normalized value = test value \\frac{t_{measured}}{t_{nominal}}}{normalized value = test value * t_measured / t_nominal}

If you need to normalize based on fiber volume fraction (or another method),
you will first need to calculate the nominal cured ply thickness (or laminate
thickness). Those calculations are outside the scope of this documentation.
}
