% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{normalize_group_mean}
\alias{normalize_group_mean}
\title{Normalize values to group means}
\usage{
normalize_group_mean(x, group)
}
\arguments{
\item{x}{the variable containing the data to normalized}

\item{group}{the variable containing the groups}
}
\value{
Returns a vector of normalized values
}
\description{
This function computes the mean of each group, then divides each group by
their corresponding group mean. This is commonly done
when pooling data across environments.
}
\details{
Computes the mean for each group, then divides each value by the mean for
the corresponding group.
}
\references{
“Composite Materials Handbook, Volume 1. Polymer Matrix Composites
Guideline for Characterization of Structural Materials,” SAE International,
CMH-17-1G, Mar. 2012.
}
