% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{print.basis}
\alias{print.basis}
\title{Nicely formats the results of basis calculations}
\usage{
\method{print}{basis}(x, ...)
}
\arguments{
\item{x}{the \code{basis} object to be printed}

\item{...}{additional arguments to be passed to \code{\link[base]{format}}}
}
\description{
Nicely formats the results of basis calculations
}
