% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{print.equiv_change_mean}
\alias{print.equiv_change_mean}
\title{Nicely formats the results of \code{\link{equiv_change_mean}}}
\usage{
\method{print}{equiv_change_mean}(x, ...)
}
\arguments{
\item{x}{the equiv_change_mean object to be printed}

\item{...}{additional arguments to be passed to \code{\link[base]{format}}}
}
\description{
Nicely formats the results of \code{\link{equiv_change_mean}}
}
