% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{print.equiv_mean_extremum}
\alias{print.equiv_mean_extremum}
\title{Nicely formats the results from \code{\link{equiv_mean_extremum}}}
\usage{
\method{print}{equiv_mean_extremum}(x, ...)
}
\arguments{
\item{x}{the equiv_mean_extremum object to be printed}

\item{...}{additional arguments to be passed to \code{\link[base]{format}}}
}
\description{
Nicely formats the results from \code{\link{equiv_mean_extremum}}
}
