% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{start_analysis}
\alias{start_analysis}
\title{Start an analysis for a data set}
\usage{
start_analysis(data, test, condition = NULL, subgroup = NULL)
}
\arguments{
\item{data}{a \code{data.frame} (or similar) containing the data to
analyze}

\item{test}{the column in the data.frame indicating the test
for each observation}

\item{condition}{(optional) the column in the data.frame indicating
the environmental condition of each observation}

\item{subgroup}{(optional) a column in the data.frame indicating
a further sub-division of the data, for example rows
in the data.frame containing normalized or as-measured
values}
}
\value{
an object of class \code{cmanalysis} that can be passed to
\code{\link{analyze_property}}.
}
\description{
Accepts a \code{data.frame} (or similar object) and starts an analysis.
To start an analysis, this function needs to know how to interpret the
columns of the \code{data.frame}: the columns representing the test,
the environmental condition (optional) and, optionally,
a further sub-division of the data. Once the analysis is started,
individual properties can be analyzed.
}
\seealso{
\code{\link{analyze_property}}
}
