% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stat_normal_surv_func}
\alias{stat_normal_surv_func}
\title{Normal Survival Function}
\usage{
stat_normal_surv_func(
  mapping = NULL,
  data = NULL,
  geom = "smooth",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  n = 100,
  pad = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}.}

\item{data}{The data to be displayed in this layer. This has the
same usage as a \code{ggplot2} \code{stat} function.}

\item{geom}{The geometric object to use to display the data.}

\item{position}{Position argument}

\item{show.legend}{Should this layer be included in the legends?}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetic,
rather than combining with them.}

\item{n}{If \code{NULL}, do not interpolated. Otherwise, the
number of points to interpolate.}

\item{pad}{If \code{TRUE}, pad the ESF with additional points
\verb{(-Inf, 0)} and \verb{(0, Inf)}.}

\item{...}{Other arguments to pass on to \code{layer}.}
}
\description{
The Normal survival function provides a visualization of a
distribution. A normal curve is fit based on the mean and standard
deviation of the data, and the survival function of this normal
curve is plotted. The survival function is simply one minus the
CDF.
}
