% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtest.R
\name{anderson_darling}
\alias{anderson_darling}
\title{Anderson-Darling test for goodness of fit}
\usage{
anderson_darling(x, dist, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{dist}{the cumulative distribution function considered}

\item{...}{additional arguments to pass to \code{dist}}
}
\value{

}
\description{
TODO: Write this
}
\details{
TODO: Write this
}
\references{
J. F. Lawless, \emph{Statistical models and methods for lifetime data}.
New York: Wiley, 1982.
}
