% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtest.R
\name{anderson_darling}
\alias{anderson_darling}
\alias{anderson_darling_normal}
\alias{anderson_darling_lognormal}
\alias{anderson_darling_weibull}
\title{Anderson-Darling test for goodness of fit}
\usage{
anderson_darling_normal(data = NULL, x, alpha = 0.05)

anderson_darling_lognormal(data = NULL, x, alpha = 0.05)

anderson_darling_weibull(data = NULL, x, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame-like object (optional)}

\item{x}{a numeric vector or a variable in the data.frame}

\item{alpha}{the required significance to conclude that the data follows
the specified distribution. Defaults to 0.05.}
}
\value{
an object of class \code{anderson_darling}. This object has the following
fields.

\item{\code{call}}{the expression used to call this function}
\item{\code{dist}}{the distribution used}
\item{\code{data}}{a copy of the data analyzed}
\item{\code{n}}{the number of observations in the sample}
\item{\code{A}}{the Anderson-Darling test statistic}
\item{\code{osl}}{the significance level, assuming the
  parameters of the distribution are estimated from the data}
\item{\code{alpha}}{the required significance level for the test to
 conclude that the data is drawn from the specified distribution.
 This value is given by the user.}
\item{\code{reject_distribution}}{a logical value indicating whether
 the hypothesis that the data is drawn from the specified distribution
 should be rejected}
}
\description{
Calculates the Anderson-Darling test statistic for a sample, given
a particular distribution and computes the significance level.
}
\details{
The Anderson-Darling test statistic is calculated for the distribution
given.

The significance level is calculated assuming that the parameters
of the distribution are unknown; these parameters are estimate from the
data.

The function \code{anderson_darling_normal} computes the Anderson-Darling
test statistic given a normal distribution with mean and standard deviation
equal to the sample mean and standard deviation.

The function \code{anderson_darling_lognormal} is the same as
\code{anderson_darling_normal} except that the data is log transformed
first.

The function \code{anderson_darling_weibull} computes the Anderson-Darling
test statistic given a Weibull distribution with shape and scale parameters
estimate from the data using a maximum likelihood estimate.

This function uses the formula for observed significance
level published in CMH-17-1G. These formulae depend on the particular
distribution used.

The results of this function have been validated against
published values in Lawless (1982).
}
\references{
J. F. Lawless, \emph{Statistical models and methods for lifetime data}.
New York: Wiley, 1982.

"Composites Materials Handbook, Volume 1. Polymer Matrix
Composites Guideline for Characterization of Structural
Materials," SAE International, CMH-17-1G, Mar. 2012.
}
