% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbon.fabric}
\alias{carbon.fabric}
\alias{carbon.fabric.2}
\title{Sample data for a generic carbon fabric}
\format{An object of class \code{data.frame} with 216 rows and 5 columns.}
\usage{
carbon.fabric

carbon.fabric.2
}
\description{
Datasets containing sample data that is typical of a generic carbon
prepreg fabric. This data is used in several examples within the
\code{cmstatr} package. This data is fictional and should
not be used for anything besides learning how to use this package.
}
\keyword{datasets}
