% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmstatr.R
\docType{package}
\name{cmstatr-package}
\alias{cmstatr}
\alias{cmstatr-package}
\title{cmstatr: Statistical Methods for Composite Material Data}
\description{
To learn more about \code{cmstatr}, start with the vignettes:
\code{browseVignettes(package = "cmstatr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.cmstatr.net/}
  \item \url{https://github.com/cmstatr/cmstatr}
  \item Report bugs at \url{https://github.com/cmstatr/cmstatr/issues}
}

}
\author{
\strong{Maintainer}: Stefan Kloppenborg \email{stefan@kloppenborg.ca}

Other contributors:
\itemize{
  \item Billy Cheng \email{bcheng@comtekadvanced.com} [contributor]
  \item Ally Fraser \email{ally.fraser25@gmail.com} [contributor]
  \item Comtek Advanced Structures, Ltd. [funder]
}

}
\keyword{internal}
