% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate the coefficient of variation}
\usage{
cv(x)
}
\arguments{
\item{x}{a vector}
}
\value{
The calculated CV
}
\description{
The coefficient of variation (CV) is the ratio of the standard
deviation to the mean of a sample. This function takes a vector
of data and calculates the CV.
}
