% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{glance.basis}
\alias{glance.basis}
\title{Glance at a basis object}
\usage{
\method{glance}{basis}(x, ...)
}
\arguments{
\item{x}{a basis object}

\item{...}{Additional arguments. Not used. Included only to match generic
signature.}
}
\value{
A one-row \code{\link[tibble:tibble]{tibble::tibble}} with the following
columns:

\item{\code{p}}{the the proportion of the population that the basis value
       should be below. Normally 0.90 or 0.99}
\item{\code{conf}}{The confidence level. Normally 0.95}
\item{\code{distribution}}{A string representing the distribution assumed
       when calculating the basis value}
\item{\code{r}}{the sample size}
\item{\code{r}}{the number of groups used in the calculation. This will
       be \code{NA} for single-point basis values}
\item{\code{basis}}{the basis value}
}
\description{
Glance accepts an object of type basis and returns a
\code{\link[tibble:tibble]{tibble::tibble}} with
one row of summaries.

Glance does not do any calculations: it just gathers the results in a
tibble.
}
\examples{
x <- rnorm(20, 100, 5)
b <- basis_normal(x = x)
glance(b)

#> ## A tibble: 1 x 6
#>      p  conf distribution  n r     basis
#>  <dbl> <dbl> <chr>     <int> <lgl> <dbl>
#>      1   0.9  0.95 Normal 20 NA    87.8

}
\seealso{
\code{\link{basis}}
}
