% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{k_equiv}
\alias{k_equiv}
\title{k-factors for equivalency testing}
\usage{
k_equiv(alpha, n)
}
\arguments{
\item{alpha}{the acceptable probability of a type I error}

\item{n}{the number of observations in the sample to test for equivalency}
}
\value{
a vector with elements c(k1, k2). k1 is for testing the sample
  extremum. k2 is for testing the sample mean
}
\description{
k-factors for equivalency testing
}
\details{
The k-factors returned by this function are used for equivalency testing,
  considering both the sample mean and sample extremum (normally minimum).
  The results of this function match those published by Vangel within
  0.05\% for \eqn{n > 2} and \eqn{\alpha \le 0.25}. Those results published
  by Vangel are identical to those published in CMH-17-1G.

  This function uses numerical integration and numerical optimization to
  find values of the factors \eqn{k_1} and \eqn{k_2} based on Vangel's
  saddle point approximation.

  The value \eqn{n} refers to the number of observations in the sample
  being compared with the original population (the qualification sample is
  usually assumed to be equal to the population statistics).

  The value of \eqn{alpha} is the acceptable probability of a type I error.
  Normally, this is set to 0.05 for material or process equivalency and 0.01
  when setting lot acceptance limits. Though, in principle, this parameter
  can be set to any number between 0 and 1. This function, however, has only
  been validated in the range of \eqn{1e-5 \le alpha \le 0.5}.
}
\examples{
qual_mean <- 100
qual_sd <- 3.5
k <- k_equiv(0.01, 5)
acceptance_limit_min <- qual_mean - qual_sd * k[1]
acceptance_limit_mean <- qual_mean - qual_sd * k[2]
}
\references{
M. G. Vangel. Lot Acceptance and Compliance Testing Using the Sample Mean
  and an Extremum, Technometrics, vol. 44, no. 3. pp. 242–249.
}
