% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene.R
\name{levene_test}
\alias{levene_test}
\title{Levene's Test for Equality of Variance}
\usage{
levene_test(data, x, groups, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame}

\item{x}{the variable in the data.frame or a vector on which to perform the
Levene's test (usually strength)}

\item{groups}{a variable in the data.frame that defines the groups}

\item{alpha}{the significance level (default 0.05)}
}
\value{
Returns an object of class \code{adk}. This object has the following fields:
\describe{
  \item{\code{call}}{the expression used to call this function}
  \item{\code{data}}{the original data used to compute the ADK}
  \item{\code{groups}}{a vector of the groups used in the computation}
  \item{\code{alpha}}{the value of alpha specified}
  \item{\code{n}}{the total number of observations}
  \item{\code{k}}{the number of groups}
  \item{\code{f}}{the value of the F test statistic}
  \item{\code{p}}{the computed p-value}
  \item{\code{reject_equal_variance}}{a boolean value indicating whether the
    null hypothesis that all samples have the same variance is rejected}
}
}
\description{
This function performs the Levene's test for equality of variance.
}
\details{
This function performs the Levene's test for equality of variance. The
data is transformed as follows:

\deqn{wij = | xij - mi |}

Where \eqn{mi} is median of the \eqn{ith} group. An F-Test is then
performed on the transformed data.
}
\references{
“Composites Materials Handbook, Volume 1. Polymer Matrix Composites
Guideline for Characterization of Structural Materials,” SAE International,
CMH-17-1G, Mar. 2012.
}
