% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnr.R
\name{maximum_normed_residual}
\alias{maximum_normed_residual}
\title{Detect outliers using the maximum normed resiudal method}
\usage{
maximum_normed_residual(df = NULL, x, alpha = 0.05)
}
\arguments{
\item{df}{a data.frame}

\item{x}{the variable in the data.frame for which to find the MNR
or a vector if \code{df=NULL}}

\item{alpha}{the significance level for the test. Defaults to 0.05}
}
\value{
an object of class \code{mnr}
This object has the following fields:
\describe{
  \item{\code{call}}{the expression used to call this function}
  \item{\code{data}}{the original data used to compute the MNR}
  \item{\code{mnr}}{the computed MNR test statistic}
  \item{\code{crit}}{the critical value given the sample size and the
                     significance level}
  \item{\code{outliers}}{a data.frame containing the \code{index} and
                         \code{value} of each of the identified outliers}
  \item{\code{n_outliers}}{the number of outliers found}
}
}
\description{
Detect outliers using the maximum normed resiudal method
}
\details{
\code{df} is an optional argument. If \code{df} is given, it should be a
\code{data.frame} (or similar object). When \code{df} is specified, the
value of \code{x} is expected to be a variable within \code{df}. If
\code{df} is not specified, \code{x} must be a vector.

The maximum normed residual test is a test for outliers. The test statistic
is given in CMH-17-1G. Outliers are identified in the returned object.
}
