% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{transform_mod_cv}
\alias{transform_mod_cv}
\title{Transforms data according to the modified CV rule}
\usage{
transform_mod_cv(x, groups = NULL)
}
\arguments{
\item{x}{a vector of data to transform}

\item{groups}{a vector indicating the group to which each observation in
\code{x} belongs. If this is NULL, the data will be treated as
unstructured (without grouping)}
}
\value{
A vector of transformed data
}
\description{
Transforms data according to the modified CV rules. Takes a vector
containing the observations and a vector containing the groups that
each observation belongs to and transforms the data. The modified CV
is computed within each group and each observation is transformed
according to:

\deqn{\frac{S^*}{S} \left(x_i - \bar{x_i}\right) + \bar{x_i}}{
  S*/S (xi-x_bar_i) + x_bar_i
}
}
\seealso{
\code{\link{calc_cv_star}}
}
