% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{transform_mod_cv}
\alias{transform_mod_cv}
\title{Transforms data according to the modified CV rule}
\usage{
transform_mod_cv(x, groups = NULL)
}
\arguments{
\item{x}{a vector of data to transform}

\item{groups}{a vector indicating the group to which each observation in
\code{x} belongs. If this is NULL, the data will be treated as
unstructured (without grouping)}
}
\value{
A vector of transformed data
}
\description{
Transforms data according to the modified CV rules. Takes a vector
containing the observations and a vector containing the groups that
each observation belongs to and transforms the data. The modified CV
is computed within each group and each observation is transformed
according to:

\deqn{\frac{S_i^*}{S_i} (x_i - \bar{x_i}) + \bar{x_i}}{
  Si*/Si (xi-x_bar_i) + x_bar_i
}

Where Si* is the modified standard deviation (mod CV times mean) for
the group; Si is the standard deviation for the group, x_bar_i is
the group mean and xi is the observation.
}
\examples{
# Transform data according to the modified CV transformation
# and report the original and modified CV for each condition

library(tidyverse)
carbon.fabric \%>\%
  filter(test == "FT") \%>\%
  mutate(trans_strength = transform_mod_cv(strength, condition)) \%>\%
  group_by(condition) \%>\%
  summarize(cv = sd(strength) / mean(strength),
            mod_cv = sd(trans_strength) / mean(trans_strength))

## # A tibble: 3 x 3
##   condition     cv mod_cv
##   <chr>      <dbl>  <dbl>
## 1 CTD       0.0423 0.0612
## 2 ETW       0.0369 0.0600
## 3 RTD       0.0621 0.0711

}
\seealso{
\code{\link{calc_cv_star}}
}
